Reads a text file and prints it to the standard output

+++ 0x15-file_io/0-read_textfile.c
@@ -11,7 +11,7 @@
   */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	ssize_t fd, rd, wr;
+	int fd, rd, wr;
 	char *buff;
 
 	if (filename == NULL)
